Cleaned up the custom.css to eliminate all the styles (the file was not included in the application). Added it to the application and added a style for legend-div class. This allows us to style the heading on the legends which I changed the font-family and font-size to more closely match the text in the image files.
Made phragmites italics everywhere it shows up, including the page header.

Added a drawingOrder property to the habitLayers array objects. The layers are then sorted by drawingOrder prior to adding to the map. This allows us to control the order of the layer drawing while still allowing the array order to control how they are displayed.

git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/glri-phragmites-map/trunk@23206 622dc1fb-2224-0410-ab72-dc85a3975e1a